fix autocomplite and rename place to location
