+++ webapp/methods.py
@@ -1,5 +1,6 @@
 from webapp.user.models import User, UserProfile
 from webapp.meeting.models import GameMeeting
+from webapp.location.models import City
 from flask_wtf import FlaskForm
 
 
@@ -25,10 +26,13 @@ def update_meeting(session, form: FlaskForm, meeting_id: int) -> None:
     """
     Обновляет данные встречи в БД
     """
+    db_city = session.query(City).filter(City.name == form['meeting_city'].data).first()
+    city_id = db_city.id if db_city else None
     meet = session.query(GameMeeting).filter(GameMeeting.id == meeting_id).first()
     meet.game_name = form['game_name'].data
     meet.number_of_players = form['number_of_players'].data
     meet.city_name = form['meeting_city'].data
+    meet.city_id = city_id,
     meet.meeting_place = form['meeting_place'].data
     meet.meeting_date_time = f"{form['date_meeting'].data} {form['time_meeting'].data}"
     meet.description = form['description'].data
