+++ webapp/templates/edit_meeting.html
@@ -36,9 +36,9 @@
             </div>
             <script>
               $(function() {
-                $("#place_autocomplete_city").autocomplete_city({
+                $("#city_autocomplete").autocomplete({
                   source:function(request, response) {
-                    $.getJSON("{{url_for('place.autocomplete_city')}}",{
+                    $.getJSON("{{url_for('location.autocomplete_city')}}",{
                       q: request.term,
                     }, function(data) {
                       response(data);
